Test of return value from server
Added two input labes in index.html for user input ('name' and 'dogName').
The server returns the phrase "Hello 'name'! Did you feed 'dogName'?".
This test is to show that the request received by the server can be handled and a response can be sent according to the user input.